Improve integer/floating-point comparisons
Also move numeric implementations to separate module.
